refactor(sketch-to-ui): simplify fuzzy attribute parser responsibilities
Extract token mapping, value resolution, and widget-specific recovery while preserving XML generation behavior.
