+++ sketch-to-ui-plugin/src/main/java/org/appdevforall/codeonthego/computervision/domain/parser/AttributeTokenMapper.kt
@@ -1,0 +1,175 @@
+package org.appdevforall.codeonthego.computervision.domain.parser
+
+import me.xdrop.fuzzywuzzy.FuzzySearch
+import org.appdevforall.codeonthego.computervision.domain.grammar.UiGrammarValidator
+import java.lang.StringBuilder
+
+internal object AttributeTokenMapper {
+    private const val PIPE_DELIMITER = "|"
+    private val inputTypeValues = InputTypeValueSet.values.map { it.lowercase() }.toSet()
+    private val grammarValidator = UiGrammarValidator()
+    private val numericTypes = setOf(
+        ValueType.DIMENSION,
+        ValueType.SP_DIMENSION,
+        ValueType.INTEGER,
+        ValueType.FLOAT
+    )
+
+    fun tokenize(annotation: String): List<String> {
+        return if (annotation.contains(PIPE_DELIMITER)) {
+            annotation.split(PIPE_DELIMITER)
+                .flatMap(::tokenizeDelimitedChunk)
+                .filter { it.isNotEmpty() }
+        } else {
+            annotation.split(Regex("[:;]|\\s+")).map { it.trim() }.filter { it.isNotEmpty() }
+        }
+    }
+
+    fun mapTokensToAttributes(tokens: List<String>, tag: String): Map<String, String> {
+        val state = MappingState()
+
+        for (token in tokens) {
+            consumeToken(token, resolveMatchedKey(token, state), tag, state)
+        }
+
+        flushCurrentAttribute(tag, state)
+        return state.attributes
+    }
+
+    fun fuzzyMatchKey(rawKey: String): AttributeKey? {
+        val normalizedKey = normalizeKeyToken(rawKey)
+
+        val exactMatch = AttributeKey.findByAlias(normalizedKey)
+        if (exactMatch != null) return exactMatch
+
+        if (normalizedKey.length <= 2) return null
+
+        val threshold = when {
+            normalizedKey.length <= 3 -> 65
+            normalizedKey.length <= 6 -> 75
+            else -> 80
+        }
+
+        val result = FuzzySearch.extractOne(normalizedKey, AttributeKey.allAliases)
+
+        return if (result.score >= threshold) AttributeKey.findByAlias(result.string) else null
+    }
+
+    fun normalizeKeyToken(rawKey: String): String {
+        return rawKey.lowercase()
+            .replace("-", "_")
+            .replace(AttributeRegexPatterns.WHITESPACE, "_")
+            .replace(".", "_")
+            .replace(AttributeRegexPatterns.MULTIPLE_UNDERSCORES, "_")
+            .replace(Regex("lay[ao0]ut"), "layout")
+            .replace(Regex("(?<=^|_)[lt]d(?=$|_)"), "id")
+    }
+
+    private fun tokenizeDelimitedChunk(chunk: String): List<String> {
+        val trimmed = chunk.trim()
+        if (trimmed.isEmpty()) return emptyList()
+
+        val delimiterIndex = trimmed.indexOfFirst { it == ':' || it == ';' }
+        if (delimiterIndex >= 0) {
+            val key = trimmed.substring(0, delimiterIndex).trim()
+            val value = trimmed.substring(delimiterIndex + 1).trim()
+            return listOf(key, value).filter { it.isNotEmpty() }
+        }
+
+        val firstSpace = trimmed.indexOfFirst { it.isWhitespace() }
+        if (firstSpace < 0) return listOf(trimmed)
+
+        val key = trimmed.substring(0, firstSpace).trim()
+        val value = trimmed.substring(firstSpace + 1).trim()
+        return listOf(key, value).filter { it.isNotEmpty() }
+    }
+
+    private fun shouldTreatTokenAsValue(
+        token: String,
+        currentKey: AttributeKey?,
+        currentValue: StringBuilder
+    ): Boolean {
+        val lowerToken = token.trim().lowercase()
+
+        return when {
+            currentKey == AttributeKey.ID &&
+                currentValue.isBlank() &&
+                AttributeKey.findByAlias(normalizeKeyToken(token)) == null -> true
+            currentKey?.valueType == ValueType.TEXT_CONTENT &&
+                currentValue.isBlank() &&
+                AttributeKey.findByAlias(normalizeKeyToken(token)) == null -> true
+            currentKey == AttributeKey.INPUT_TYPE && lowerToken in inputTypeValues -> true
+            currentKey in setOf(AttributeKey.LAYOUT_GRAVITY, AttributeKey.GRAVITY) && lowerToken in GravityValueSet.values -> true
+            currentKey?.valueType == ValueType.COLOR && isColorToken(lowerToken) -> true
+            currentKey?.valueType == ValueType.DIMENSION && DimensionValueSet.allKeywords.any { it in lowerToken } -> true
+            currentKey?.valueType in numericTypes -> lowerToken.any { it.isDigit() }
+            else -> false
+        }
+    }
+
+    private fun isColorToken(token: String): Boolean {
+        return token.startsWith("#") || token.startsWith("@") || token in ColorCleaner.colorMap
+    }
+
+    private fun resolveMatchedKey(token: String, state: MappingState): AttributeKey? {
+        if (shouldTreatTokenAsValue(token, state.currentKey, state.currentValue)) return null
+        return fuzzyMatchKey(token)
+    }
+
+    private fun consumeToken(token: String, matchedKey: AttributeKey?, tag: String, state: MappingState) {
+        if (matchedKey == null) {
+            appendTokenValue(token, state)
+            return
+        }
+
+        startNewAttribute(matchedKey, tag, state)
+    }
+
+    private fun startNewAttribute(matchedKey: AttributeKey, tag: String, state: MappingState) {
+        flushCurrentAttribute(tag, state)
+        state.currentKey = matchedKey
+        state.currentValue.clear()
+    }
+
+    private fun appendTokenValue(token: String, state: MappingState) {
+        state.currentValue.append(token).append(" ")
+    }
+
+    private fun flushCurrentAttribute(tag: String, state: MappingState) {
+        flushAttribute(state.currentKey, state.currentValue.toString(), tag, state.attributes)
+    }
+
+    private fun flushAttribute(key: AttributeKey?, rawValue: String, tag: String, destination: MutableMap<String, String>) {
+        if (key == null || rawValue.isBlank()) return
+
+        val resolved = AttributeValueResolver.resolve(key, rawValue, tag) ?: return
+        val existingValue = destination[resolved.xmlAttr]
+        if (existingValue == null || shouldReplaceAttribute(resolved.xmlAttr, existingValue, resolved.value, tag)) {
+            destination[resolved.xmlAttr] = resolved.value
+        }
+    }
+
+    private fun shouldReplaceAttribute(
+        xmlAttr: String,
+        existingValue: String,
+        candidateValue: String,
+        tag: String
+    ): Boolean {
+        if (xmlAttr == AttributeKey.LAYOUT_GRAVITY.xmlName &&
+            existingValue == "center" &&
+            candidateValue == "center_horizontal"
+        ) {
+            return true
+        }
+
+        val existingValid = grammarValidator.enforceGrammar(mapOf(xmlAttr to existingValue), tag).containsKey(xmlAttr)
+        val candidateValid = grammarValidator.enforceGrammar(mapOf(xmlAttr to candidateValue), tag).containsKey(xmlAttr)
+        return !existingValid && candidateValid
+    }
+
+    private data class MappingState(
+        var currentKey: AttributeKey? = null,
+        val currentValue: StringBuilder = StringBuilder(),
+        val attributes: MutableMap<String, String> = mutableMapOf()
+    )
+}
