+++ sketch-to-ui-plugin/src/main/java/org/appdevforall/codeonthego/computervision/domain/parser/AttributeValueResolver.kt
@@ -1,0 +1,56 @@
+package org.appdevforall.codeonthego.computervision.domain.parser
+
+internal data class ResolvedAttribute(val xmlAttr: String, val value: String)
+
+internal object AttributeValueResolver {
+    private const val BUTTON_TAG = "Button"
+    private const val SWITCH_TAG = "Switch"
+
+    private val cleaners = mapOf(
+        ValueType.TEXT_CONTENT to TextContentCleaner,
+        ValueType.DIMENSION to DimensionCleaner,
+        ValueType.SP_DIMENSION to SpDimensionCleaner,
+        ValueType.COLOR to ColorCleaner,
+        ValueType.ID to IdCleaner,
+        ValueType.DRAWABLE to DrawableCleaner,
+        ValueType.INTEGER to NumberCleaner,
+        ValueType.FLOAT to FloatCleaner,
+        ValueType.TEXT_STYLE to TextStyleCleaner,
+        ValueType.RAW to ValueCleaner { it }
+    )
+
+    fun resolve(key: AttributeKey, rawValue: String, tag: String): ResolvedAttribute? {
+        val trimmedRawValue = rawValue.trim()
+        val cleaner = cleaners[key.valueType] ?: ValueCleaner { it }
+        val cleanedValue = when (key) {
+            AttributeKey.ID -> IdCleaner.clean(trimmedRawValue, tag)
+            AttributeKey.INPUT_TYPE -> InputTypeCleaner.clean(trimmedRawValue)
+            AttributeKey.LAYOUT_GRAVITY, AttributeKey.GRAVITY -> GravityCleaner.clean(trimmedRawValue)
+            else -> cleaner.clean(trimmedRawValue)
+        }
+
+        if (cleanedValue.isEmpty()) return null
+
+        val recoveredValue = recoverSwitchWidth(key, rawValue, cleanedValue, tag)
+        return resolveXmlAttribute(key, recoveredValue, tag)
+    }
+
+    private fun recoverSwitchWidth(key: AttributeKey, rawValue: String, cleanedValue: String, tag: String): String {
+        val widthLostLeadingOne = tag == SWITCH_TAG &&
+            key == AttributeKey.WIDTH &&
+            cleanedValue == "0dp" &&
+            Regex("^0{2}\\s*dp$", RegexOption.IGNORE_CASE).matches(rawValue.trim())
+
+        return if (widthLostLeadingOne) "100dp" else cleanedValue
+    }
+
+    private fun resolveXmlAttribute(key: AttributeKey, value: String, tag: String): ResolvedAttribute {
+        if (key == AttributeKey.BACKGROUND && tag == BUTTON_TAG) {
+            return ResolvedAttribute("app:backgroundTint", value)
+        }
+        if (key == AttributeKey.ID) {
+            return ResolvedAttribute(key.xmlName, value.replace(" ", "_"))
+        }
+        return ResolvedAttribute(key.xmlName, value)
+    }
+}
