+++ sketch-to-ui-plugin/src/main/java/org/appdevforall/codeonthego/computervision/domain/parser/sanitizer/OcrSanitizerFactory.kt
@@ -5,6 +5,8 @@ object OcrSanitizerFactory {
     fun createDefaultSanitizer(): OcrSanitizer {
         return CompositeOcrSanitizer(
             listOf(
+                AttributeKeyPhraseSanitizer(),
+                CompactDimensionSanitizer(),
                 ColorSanitizer(),
                 TextAttributeSanitizer(),
                 DimensionSanitizer(),
